refactor(loyalty): clear the code-health findings on the parity change
Five findings, fixed at the shape each set of siblings shared rather than one
method at a time.

queries.go carried two unrelated APIs: a per-viewer ledger with guarded
arithmetic, and broadcaster-authored counters with their own name and scope
grammar. They shared only the ent client, which is what Low Cohesion was
reporting. The counter half moves to counters.go.

BalanceTransfer was four phases in one body. Split into transferParties
(resolve and validate both sides), moveBalance (the guarded debit plus credit),
and transferOutcome (re-read after commit).

All three chat commands opened with the same two branches before their own
logic; loyaltyRun now owns that preamble. The set/add/remove verbs each
repeated the moderator check and the capability check around a one-line write,
so they become a pointsGrant table and the routing moves off Loyalty() into
pointsRun. leaderboardShow keeps only its control flow, with the argument parse
and the line building beside it.

readLoyalty projected nine numeric fields with an inline coercion each; the
fields are now a list and the coercion is one function. publicModules built two
entry shapes inline, now catalogEntry and builtinEntry over a shared
activeModule lookup.

Behaviour is unchanged throughout: no verb, gate, default or wire field moved.
